+++ src/main/java/io/resume/make/domain/auth/service/KakaoOAuthService.java
@@ -1,0 +1,108 @@
+package io.resume.make.domain.auth.service;
+
+import io.resume.make.domain.auth.dto.KakaoTokenResponse;
+import io.resume.make.domain.user.repository.UserRepository;
+import io.resume.make.global.exception.BusinessException;
+import io.resume.make.global.response.GlobalErrorCode;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.core.ParameterizedTypeReference;
+import org.springframework.http.HttpHeaders;
+import org.springframework.stereotype.Service;
+import org.springframework.util.LinkedMultiValueMap;
+import org.springframework.util.MultiValueMap;
+import org.springframework.web.reactive.function.BodyInserters;
+import org.springframework.web.reactive.function.client.WebClient;
+import org.springframework.web.util.UriComponentsBuilder;
+
+import java.util.List;
+import java.util.Map;
+import java.util.UUID;
+
+@Slf4j
+@Service
+@RequiredArgsConstructor
+public class KakaoOAuthService {
+
+    private final UserRepository userRepository;
+    private final WebClient webClient;
+
+
+    @Value("${kakao.login.api_key}")
+    private String apiKey;
+
+    @Value("${kakao.login.client_secret}")
+    private String clientSecret;
+
+    @Value("${kakao.login.uri.base}")
+    private String kakaoBaseUri;
+
+    @Value("${oauth.kakao.redirect-uris}")
+    private List<String> allowedRedirectUris;
+
+    /**
+     * 카카오 로그인 url 생성
+     *
+     * @return
+     */
+    public Map<String, String> getKakaoUrl(String redirectUri, String codeChallenge) {
+        String state = UUID.randomUUID().toString();
+        log.info("State: {}", state);
+
+        if (!allowedRedirectUris.contains(redirectUri)) {
+            throw new BusinessException(GlobalErrorCode.INVALID_REDIRECT_URI);
+        }
+
+        String kakaoAuthUrl = UriComponentsBuilder.fromUriString(kakaoBaseUri)
+                .path("/oauth/authorize")
+                .queryParam("client_id", apiKey)
+                .queryParam("redirect_uri", redirectUri)
+                .queryParam("response_type", "code")
+                .queryParam("state", state)
+                .queryParam("code_challenge", codeChallenge)
+                .queryParam("code_challenge_method", "S256")
+                .build().toUriString();
+        return Map.of(
+                "loginUrl", kakaoAuthUrl,
+                "state", state
+        );
+    }
+
+    public KakaoTokenResponse exchangeKakaoToken(String code, String codeVerifier, String redirectUri) {
+        String tokenUrl = UriComponentsBuilder.fromUriString(kakaoBaseUri)
+                .path("/oauth/token")
+                .build()
+                .toUriString();
+
+        MultiValueMap<String, String> params = new LinkedMultiValueMap<>();
+        params.add("grant_type", "authorization_code");
+        params.add("client_id", apiKey);
+        params.add("redirect_uri", redirectUri);
+        params.add("code", code);
+        params.add("client_secret", clientSecret);
+        params.add("code_verifier", codeVerifier);
+
+        return webClient.post()
+                .uri(tokenUrl)
+                .header("Content-Type", "application/x-www-form-urlencoded")
+                .body(BodyInserters.fromFormData(params))
+                .retrieve().bodyToMono(KakaoTokenResponse.class)
+                .block();
+    }
+
+    public Map<String, Object> getUserInfo(String accessToken) {
+        return webClient.get()
+                .uri("http://kapi.kakao.com/v2/user/me")
+                .header(HttpHeaders.AUTHORIZATION, "Bearer " + accessToken)
+                .header(HttpHeaders.CONTENT_TYPE, "application/x-www-form-urlencoded;charset=utf-8")
+                .retrieve()
+                .bodyToMono(new ParameterizedTypeReference<Map<String, Object>>() {
+                })
+                .block();
+    }
+    /**
+     * 로그아웃
+     */
+    //TODO
+}
